Change C_drainage to lowercase, and use hard-coded default
Having this upper-case didn't work: the CESM scripts change it to
lower-case, but CISM's config file reader is case-sensitive. So Bill
Lipscomb has changed it to lowercase in CISM, and I have made the
similar change here.

Also, stop setting it explicitly in cism.config by default: Bill L
suggests using the hard-coded default for this.


git-svn-id: https://svn-ccsm-models.cgd.ucar.edu/glc/branches/develop_update_2016_07@87541 fe37f545-8307-0410-aea5-b40df96820b5