Add a test that checks compatibility with a hybrid refcase
The purpose of this testmod is to check compatibility with the hybrid
refcase that is used for out-of-the-box B1850 cases at f09_g17
resolution. Changes in CISM often break compatibility with the CISM
restart file in this refcase, so I want a test that will catch this
problem before making a CISM tag.