Rename glimmer-cism directory to source_cism
We no longer refer to CISM as glimmer-cism, so this directory name was
confusing.

Also remove other references to "glimmer".

I have *not* fixed references in test/unit/time_management: That
directory seems to need a larger overhaul if we want to maintain
it. Probably we'll just ditch it at some point.

Fixes ESCOMP/cism-wrapper#10