Edits to Section 2
A few minor changes to a very well written and helpful section.

Note: I'm trying to be consistent in referring to dycores and interfaces
as Glissade, Glide, Glad, etc., with the first letter capitalized.
All caps seem unnecessary, since most of us have forgotten the acronyms.
And with no caps, it's possible to confuse these names with ordinary words.

Example of correct usage: "I'm glad we have Glad"