Edits to Section 5
As pointed out at the top, this section needs a lot of changes.
I updated a few statements that were obsolete, but ran out of steam
when I got to the glacier_mec discussion, which is way out of date.

I didn't try to add a description of the mass-conserving SMB interpolation,
because I thought it was better to do this in conjunction with a general
rewrite.