Modification to run CISM stand alone for MISMIP+
1/ Modification of build_namelist:
Right now if a CISM variable is listed in this file it has to be present in the configuration file otherwise CESM will return an error
while building the code. For this reason I commented out some variables that are not necessary for the MISMIP+ experiment.

2/ namelist_defaults_cism.xml
This is a new file that does not contain the default cism variable used by default in cesm. That file was renamed temporarily as
namelist_defaults_cism_cesm and is not used right now.

3/ namelist_definition_cism.xml
I added the definition of the new configuration options needed for the experiment.

4/ config_compsets.xml
Create a new compset called TMISMIP which is stub everything besides CISM. The name is chosen to know on which sepcific experiment
we are working on. It could be used for other ones but in the future it is not recommended to run CISM as a stand alone model within
the cesm framework.

5/ usermods_dir
They are meant to know which modifications are needed for 2 specific experiments for MISMIP+. We can get rid of them most likely when
we work on MISOMIP.

6/ glc_comp_mct.F90
I forgot what I did in there for now.

7/ glc_import_export.F90
Typically, air temperature and smb are sent to cism via clm and the coupler. I had to comment out the vector call to CLM. This is a good
work around for now as temperature and smb are constant in MISMIP+ and we could define these fields there.
Note: It might be possible to not define anything there and still have CISM work properly as these fields are defined in the initialization
file. Maybe needs to be tried.