Adding the thickness to Export from CISM
Adding the functionality to export ice thickness from CISM to the coupler.
Note: At this point it is not clear whether thck needs to be a pointer in glc_import_export.F90. Also, I am not clear if thickness needs to be adjusted similarly to topo in glc_overid_frac.F90.