Get documentation build working with sphinx_rtd_theme
The files in _static and _templates were pulled in directly from
https://github.com/marbl-ecosys/MARBL/commit/ddd346f138b62c8415a0671f59777897ca956c99,
without any changes. The changes in conf.py were based on the changes in
that MARBL commit.

A few notes about what seems to be needed:

- It seems that the version and/or release (probably version) listed at
  the top of the file needs to match the string used in
  html_theme_options['versions']

- It seems that we really do need to list every version in
  html_theme_options['versions'] in order for the drop-down menus to be
  created correctly

- This seemed to work regardless of whether I had a versions.json file
  in the build directory

Update 2018-12-08: The above notes are no longer true now that I have
the javascript working correctly (with correct paths for our
documentation build), as I fixed in a later commit on this branch.