Cleaning up and allowing 3D field passed to cism
The main changes here are the creation of 2 3-D field to store the POP ocean temperature and salinity which are then passed on to glad.

Note: at this point, I hard coded the value of ocean layer as the z-dimension due to pressing collaboration timeline. The hard coded value (currently 7) is present in glc_fields.F90.