CISM-wrapper version for CESM tutorial 2019 (CISM)
This branch of CISM-wrapper was created to be able to run CISM stand alone within the CESM framework.
The goal is to be able to run the initMIP Greenland experiment.
This could be useful in the future but some minor issues would need to be fixed before cheking it into master.
For instence, smb and artm are not being written to the output netcdf variables while being read by CISM.
This might be due to the fact that these fields are typically sent to CISM via CLM.
That said it does not impact the result of the specific initMIP simulation.