Change NCK nuopc test to MCC
According to Jim Edwards, NCK tests are not expected to work with nuopc,
and indeed I found that there were roundoff-level differences between
the single and multi-instance versions of that test (maybe due to a
difference in number of processors?). Jim Edwards suggested changing
this to MCC, so I have done this here; this MCC test passes.