Remove T1850 noevolve test with nuopc
This test fails, but I don't think it's very important: I don't think
this is a configuration that will actually be run for scientific
purposes. I have used it for testing purposes to test the noevolve
logic, but I don't think it's very important for that: we can test this
via an I compset. (And in any case, noevolve will be a less important
configuration once we have a data glc model to serve that purpose.)
See https://github.com/ESCOMP/CMEPS/issues/142 for more comments.