Go back to having a single nmlgen with all groups
To avoid errors with the user_nl_cism file, nmlgen needs to contain all
of the groups from both cism_in and cism.config. However, for the sake
of the cism_in file, we can avoid doing the work of adding the defaults
for the cism.config groups (and vice versa).