Persist the DistGrid for each ice sheet
I'm not sure about the details of this change. However: The
ESMF_DistGridDestroy call caused the model to crash in the call to
ESMF_FieldCreate in fldlist_realize. My guess is that the mesh created
from the DistGrid (which is used in the call to ESMF_FieldCreate)
contains a reference to that DistGrid, and when that DistGrid is
destroyed, that reference becomes invalid.

Given that behavior, it seemed possible that we need to hang on to
multiple DistGrids (one per ice sheet), at least temporarily, to avoid
the DistGrid for ice sheet 2 clobbering the DistGrid for ice sheet 1.

I'm not sure when, if at all, we should then call ESMF_DistGridDestroy.