Move CISM_USE_* variables to env_run
This is needed if we want to be able to change them in a test based on
SystemTestsCompareTwo: the testing infrastructure uses a clone, which
means that variables in env_case can't be changed between the two cases.

I'm a little hesitant about this change, because it would be a challenge
for a user to change this variable after creating the case, but in
principle I think it should work, so I'm just adding a caveat in the
documentation.