Move IRT test to izumi
I wanted this test to be on cheyenne_gnu. However, IRT tests are
currently failing in the GENERATE phase (and presumably would fail in
the COMPARE phase, too). So for now I'm putting this test on izumi_nag,
since we don't do baseline generation or comparison on izumi. Once
ESMCI/cime#3978 is resolved, probably move this back to
cheyenne_gnu (though it isn't really critical for baseline generation
and comparison to be done with this test).