Increasing grid tolerance
With the new GrIS grid, the tolerance between the ESMF mesh file and the
lat-lon value in the CISM input file had a difference of about 1.15e-05
which is higher than the default tolerance of 1e-05, leading to an error
and the abort of the simulation. The tolerance was raised to 2e-05 to
circumvent the problem.